Made the example to load CSV better.
Added CSVPredict example. ConsoleExemples csvPredict (name of file to
load) , then it runs as your market predict (just loads a file instead
of yahoo)…All the example is there , build , train , evaluate , and
pruning, but prune doesn't seem to do anything.  the CSVFormLoader is
still in there, but CSVFile makes it quicker.